requirements.txt
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-allauth==0.58.2
